re-structured modules: moved format and ctrl_chars to utils
