+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod app;
 pub mod dialogs;
 pub mod ui;
+pub mod util;
 pub mod widgets;
