+++ src/widgets/commit_view.rs
@@ -1,5 +1,5 @@
 use crate::app::DiffType;
-use crate::widgets::ctrl_chars::CtrlChars;
+use crate::util::ctrl_chars::CtrlChars;
 use crate::widgets::files_view::StatefulList;
 use git2::Oid;
 use tui::buffer::Buffer;
