+++ src/widgets/mod.rs
@@ -1,7 +1,5 @@
 pub mod commit_view;
-pub mod ctrl_chars;
 pub mod diff_view;
 pub mod files_view;
-pub mod format;
 pub mod graph_view;
 pub mod models_view;
